requirements.txt
django==4.2.3
djangorestframework==3.14.0
drf-yasg==1.21.6
Pillow==10.0.0